fix: extraction integration
